refactor(memory): split migration responsibilities
